fix(common): etcd-registry fix no keepalive error

+++ common/util/mem_status_linux.go
@@ -6,9 +6,9 @@ import (
 )
 
 type MemStatus struct {
-	All  uint32 `json:"all"`
-	Used uint32 `json:"used"`
-	Free uint32 `json:"free"`
+	All  uint64 `json:"all"`
+	Used uint64 `json:"used"`
+	Free uint64 `json:"free"`
 	Self uint64 `json:"self"`
 }
 
@@ -24,8 +24,8 @@ func MemStat() MemStatus {
 	sysInfo := new(syscall.Sysinfo_t)
 	err := syscall.Sysinfo(sysInfo)
 	if err == nil {
-		mem.All = sysInfo.Totalram * uint32(syscall.Getpagesize())
-		mem.Free = sysInfo.Freeram * uint32(syscall.Getpagesize())
+		mem.All = sysInfo.Totalram * uint64(syscall.Getpagesize())
+		mem.Free = sysInfo.Freeram * uint64(syscall.Getpagesize())
 		mem.Used = mem.All - mem.Free
 	}
 	return mem
